+++ Sources/UI/Receipts/ReceiptsListScreen.swift
@@ -9,6 +9,7 @@ import SwiftUI
 import SnabbleCore
 import Combine
 
+
 public extension PurchaseProviding {
     var dateString: String? {
         let dateFormatter = DateFormatter()
@@ -70,8 +71,10 @@ public struct ReceiptsItemView: View {
                         .font(.footnote)
             }
             Spacer()
-            Text(provider.amount)
-                .font(.footnote)
+            if let amount = provider.amount {
+                Text(amount)
+                    .font(.footnote)
+            }
             Image(systemName: "chevron.right")
                 .font(.footnote)
                 .foregroundColor(.secondary)
@@ -112,10 +115,13 @@ public struct ReceiptsListScreen: View {
         AsyncContentView(source: viewModel) { output in
             VStack {
                 List {
+                    if viewModel.awaitingReceipts {
+                        Text(Asset.localizedString(forKey: "Home.Grab.Purchase.placeholder"))
+                            .multilineTextAlignment(.center)
+                            .foregroundColor(.secondary)
+                    }
                     ForEach(output, id: \.id) { provider in
                         ReceiptsItemView(provider: provider, image: viewModel.imageFor(projectId: provider.projectId))
-                        // provide a modifier for unloaded receipts here like:
-                        // .foregroundColor(provider.unloaded ? .primary : .secondary)
                             .onTapGesture {
                                 viewModel.actionPublisher.send(provider)
                             }